move assets stuf to own file
